Identity Parcial

+++ Commerce/Data/Mappings/EntregaMapping.cs
@@ -9,7 +9,7 @@ namespace Commerce.Data.Mappings
         public void Configure(EntityTypeBuilder<Entrega> builder)
         {
             builder.HasKey(p => p.Id);
-            builder.Property(p => p.PrazoDias).IsRequired().HasColumnType("int");
+            builder.Property(p => p.PrazoDias).IsRequired().HasColumnType("bigint");
             builder.Property(p => p.ValorFrete).IsRequired().HasColumnType("DECIMAL(11,2)");
             builder.HasOne(p => p.Pedido).WithOne(p => p.Entrega);
             // builder.HasOne<Pedido>(p => p.Pedido).WithOne(p => p.Entrega).HasForeignKey<Entrega>(p => p.PedidoId);
